requirements.txt
# CogRepo Requirements

# Core Dependencies (for import and enrichment)
anthropic>=0.26.0          # Claude API for AI enrichment
python-dateutil>=2.8.2     # Flexible date parsing
pyyaml>=6.0                # Configuration file parsing
python-dotenv>=1.0.0       # Load environment variables from .env file

# Web UI Dependencies
flask>=3.0.0               # Web framework for UI server
flask-socketio>=5.3.0      # WebSocket support for real-time progress
flask-cors>=4.0.0          # CORS support for API
werkzeug>=3.0.0            # File upload handling

# Production Server
gunicorn>=21.0.0           # WSGI server for production
eventlet>=0.33.0           # Async support for WebSockets
psutil>=5.9.0              # System monitoring for health checks

# Configuration & Validation
pydantic>=2.0.0            # Data validation
pydantic-settings>=2.0.0   # Settings management

# Optional Dependencies (for advanced features)
# beautifulsoup4>=4.12.0   # HTML parsing for Gemini Docs exports (optional)
# lxml>=5.2.0              # XML/HTML parsing backend (optional)
pandas>=2.2.0              # Parquet file generation
pyarrow>=15.0.0            # Parquet backend

# Intelligence Layer (v3 Features)
sentence-transformers>=2.2.2  # Embedding generation for semantic search
scikit-learn>=1.3.0           # Clustering and dimensionality reduction
hdbscan>=0.8.33               # Hierarchical density-based clustering
numpy>=1.24.0                 # Numerical operations for embeddings
scipy>=1.11.0                 # Scientific computing for similarity metrics

# SOTA Enrichment (v4 Features)
openai>=1.0.0                 # OpenAI API for text-embedding-3-large

# CLI and Progress
tqdm>=4.66.0               # Progress bars for batch processing

# Reliability
tenacity>=8.2.3            # Retry logic with exponential backoff

# Testing
pytest>=8.1.0              # Testing framework
pytest-mock>=3.14.0        # Mocking for tests

# Python Standard Library (no install needed):
# - json, datetime, pathlib, sys, collections
# - http.server, urllib.parse, math, os, re
# - hashlib, dataclasses, abc, typing